Wire up prek hooks: ruff, mypy, codespell, bandit, uv-lock, formatters
Adds a `prek.toml` adapted from the givenergy-modbus repo, with the
poetry hooks swapped for `astral-sh/uv-pre-commit` (uv-lock).

Project-specific scoping:
- absolufy-imports: excluded for `custom_components/` — HA loads the
  integration as `givenergy_local`, not `custom_components.givenergy_local`,
  so intra-package imports must stay relative; tests still get processed.
- codespell: ignore `hass` (HA fixture) and `astroid` (real lib in uv.lock).
- bandit: skip B101 (asserts here are type-narrowing, not security guards).

Mypy config in pyproject.toml:
- `explicit_package_bases` + `namespace_packages` + `files = custom_components`
  to resolve the `custom_components.X` vs `X` duplicate-module ambiguity.
- Disabled three strict sub-flags (`disallow_subclassing_any`,
  `disallow_any_generics`, `warn_return_any`) since HA core ships untyped
  and givenergy-modbus's `Inverter`/`Battery` are built via pydantic
  `create_model`, surfacing as `Any` regardless of `py.typed`.
- `tests.*` override: `ignore_errors = true`.

Real type errors flushed out by mypy:
- coordinator helpers: `assert self._client is not None` for mypy's benefit
  (the contract is enforced upstream in `_async_update_data`).
- number.py: `set_value_cmd` typed `Callable[[float], list]` to match how
  HA invokes `async_set_native_value`.
- config_flow.py: `# type: ignore[call-arg]` on the HA `domain=DOMAIN`
  class-kwarg idiom.

Ruff `line-length` bumped to 100 (matches HA core); reformatting applied.

+++ custom_components/givenergy_local/number.py
@@ -1,8 +1,7 @@
 from __future__ import annotations
 
+from collections.abc import Callable
 from dataclasses import dataclass, field
-from collections.abc import Callable, Awaitable
-from typing import Any
 
 from givenergy_modbus.client import commands
 from givenergy_modbus.model.inverter import Inverter
@@ -21,7 +20,7 @@ from .coordinator import GivEnergyUpdateCoordinator
 @dataclass(frozen=True, kw_only=True)
 class GivEnergyNumberEntityDescription(NumberEntityDescription):
     value_fn: Callable[[Inverter], float | None] = field(default=lambda _: None)
-    set_value_cmd: Callable[[int], list] = field(default=lambda _: [])
+    set_value_cmd: Callable[[float], list] = field(default=lambda _: [])
 
 
 NUMBER_DESCRIPTIONS: tuple[GivEnergyNumberEntityDescription, ...] = (
@@ -95,8 +94,7 @@ async def async_setup_entry(
 ) -> None:
     coordinator: GivEnergyUpdateCoordinator = hass.data[DOMAIN][entry.entry_id]
     async_add_entities(
-        GivEnergyNumberEntity(coordinator, description)
-        for description in NUMBER_DESCRIPTIONS
+        GivEnergyNumberEntity(coordinator, description) for description in NUMBER_DESCRIPTIONS
     )
